chardev: Converting public IO impls to size_t
Due to an integer overflow issue found in ccid_card_vscard_read
(CVE-2018-18438), memory was previously susceptible to corruption. To
mitigate this issue, a number of components, including the
ccid-card-passthrough device are now utilizing size_t in place of int to
prevent an the overflow from occuring when values larger than
available space are passed to corresponding functions.

While CCID smart card readers were the primary target of the
vulnerability, a number of other subsystems have also been modified in the
same manner to prevent them from being exploited in the same way; specifically
functions which are modeled after the IOCanReadHandler and IOReadHandler
typedefs.

Subsystems affected include, but are not limited to:
CCID Smart Cards
Spice consoles
Monitors
Network Sockets
Network Tap Devices
Serial IO
Emulated Shell Serial IO
USB Redirection and Passthrough
Consoles
SLiRP IO
Mouse Emulation
Xen Consoles
SLCP IO
Network Block Devices

(cherry picked from commit 4792016a6119a4d5768bd8f989ae90b52dfef789)
[LD: BSC#1112185 CVE-2018-18438]

+++ backends/rng-egd.c
@@ -45,11 +45,11 @@ static void rng_egd_request_entropy(RngBackend *b, RngRequest *req)
     }
 }
 
-static int rng_egd_chr_can_read(void *opaque)
+static size_t rng_egd_chr_can_read(void *opaque)
 {
     RngEgd *s = RNG_EGD(opaque);
     GSList *i;
-    int size = 0;
+    size_t size = 0;
 
     for (i = s->parent.requests; i; i = i->next) {
         RngRequest *req = i->data;
@@ -59,7 +59,7 @@ static int rng_egd_chr_can_read(void *opaque)
     return size;
 }
 
-static void rng_egd_chr_read(void *opaque, const uint8_t *buf, int size)
+static void rng_egd_chr_read(void *opaque, const uint8_t *buf, size_t size)
 {
     RngEgd *s = RNG_EGD(opaque);
