+++ hw/usb/dev-serial.c
@@ -417,7 +417,7 @@ static void usb_serial_handle_destroy(USBDevice *dev)
     qemu_chr_add_handlers(s->cs, NULL, NULL, NULL, NULL);
 }
 
-static int usb_serial_can_read(void *opaque)
+static size_t usb_serial_can_read(void *opaque)
 {
     USBSerialState *s = opaque;
 
@@ -427,7 +427,7 @@ static int usb_serial_can_read(void *opaque)
     return RECV_BUF - s->recv_used;
 }
 
-static void usb_serial_read(void *opaque, const uint8_t *buf, int size)
+static void usb_serial_read(void *opaque, const uint8_t *buf, size_t size)
 {
     USBSerialState *s = opaque;
     int first_size, start;
